Qibotn integration to Qibo

+++ src/qibotn/circuit_to_mps.py
@@ -2,8 +2,8 @@ import cupy as cp
 import numpy as np
 from cuquantum import cutensornet as cutn
 
-from qibotn.MPSUtils import apply_gate, initial
-from qibotn.QiboCircuitConvertor import QiboCircuitToEinsum
+from qibotn.circuit_convertor import QiboCircuitToEinsum
+from qibotn.mps_utils import apply_gate, initial
 
 
 class QiboCircuitToMPS:
@@ -21,7 +21,7 @@ class QiboCircuitToMPS:
         self.handle = cutn.create()
         self.dtype = dtype
         self.mps_tensors = initial(self.num_qubits, dtype=dtype)
-        circuitconvertor = QiboCircuitToEinsum(circ_qibo)
+        circuitconvertor = QiboCircuitToEinsum(circ_qibo, dtype=dtype)
 
         for gate, qubits in circuitconvertor.gate_tensors:
             # mapping from qubits to qubit indices
